Broke Pages into Components
