+++ app/restaurant/[slug]/components/Description.tsx
@@ -1,0 +1,14 @@
+
+export default function Description() {
+  return (
+    <div className="mt-4">
+      <p className="text-lg font-light">
+        The classics you love prepared with a perfect twist, all served up in an
+        atmosphere that feels just right. Thats the Milestones promise. So,
+        whether youre celebrating a milestone, making the most of Happy Hour or
+        enjoying brunch with friends, you can be sure that every Milestones
+        experience is a simple and perfectly memorable one.
+      </p>
+    </div>
+  );
+}
